Consider new Turbolinks 5 event names
Turbolinks 5 has renamed their events, so we should rely on the
`turbolinks:load` to make sweet alerts work.

I'm adding this event but not removing `page:load` and `page:update`
in order to support the classic version of Turbolinks
(https://github.com/turbolinks/turbolinks-classic)

In this way, rails apps using either version of TL are covered.